+++ tsconfig.cjs
@@ -1,0 +1,21 @@
+const { ignorePaths } = require('./cspell.json');
+
+module.exports = {
+  compilerOptions: {
+    target: 'ES2022',
+    module: 'NodeNext',
+    moduleResolution: 'NodeNext',
+    strict: true,
+    skipLibCheck: true,
+    noEmit: false,
+    outDir: './dist',
+    rootDir: './',
+    declaration: true,
+    declarationMap: true,
+    sourceMap: true
+  },
+  include: [
+    '**/*.ts',
+  ],
+  exclude: [...ignorePaths]
+};